refactor: replace page components
