+++ reactjs-films-homework/src/pages/ActorProfile/ActorProfile.tsx
