+++ reactjs-films-homework/src/pages/Main/Main.tsx
