+++ reactjs-films-homework/src/pages/MovieDetails/MovieDetails.tsx
