+++ reactjs-films-homework/src/pages/index.tsx
@@ -1,6 +1,6 @@
-import { ActorProfile } from "./ActorProfile";
-import { Main } from "./Main";
-import { MovieDetails } from "./MovieDetails";
-import { NotFoundPage } from "./NotFoundPage";
+import { ActorProfile } from "./ActorProfile/ActorProfile";
+import { Main } from "./Main/Main";
+import { MovieDetails } from "./MovieDetails/MovieDetails";
+import { NotFoundPage } from "./NotFoundPage/NotFoundPage";
 
 export { ActorProfile, Main, MovieDetails, NotFoundPage };
